backend/requirements.txt
# ── Production dependencies ────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.21
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
greenlet==3.1.1

# Authentication
PyJWT==2.10.1
bcrypt==4.2.1
email-validator==2.2.0
